refactor(web): unify the two model detail modals
Dashboard and Arena opened a bare ModelDetailPanel in a Modal wrapper -
a second, plainer model-detail surface that missed every styling pass
and had no usage snippets. The rich Models-page modal now accepts its
six management callbacks as optional: when absent (read-only viewers)
the action footer and field editing are hidden, and everything else
(snippets, capabilities, pricing, modalities) comes along for free.
Dashboard and Arena use it directly; the thin wrapper in
ModelDetailPanel.tsx is gone. ModelDetailPanel itself stays as the
inline Chat side panel.
